Подключил вывод через шаблоны и экранировал пользовательские данные
